requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
fastapi-utils==0.2.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
imap-tools==0.57.0
passlib==1.7.4
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
python-multipart==0.0.5
redbird==0.5.1
rocketry==2.4.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
